refactor: HandlerMappingRegistry에서 HandlerMapping를 등록해서 사용하도록 변경
